\name{ideas-package}
\alias{ideas-package}
\docType{package}
\title{
\packageTitle{ideas}
}
\description{
\packageDescription{ideas}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ideas}
\packageIndices{ideas}
This package is for differential expression analysis of scRNA-seq data collected in multiple individuals. 
}
\author{
\packageAuthor{ideas}

Maintainer: \packageMaintainer{ideas}
}
\references{
Mengqi Zhang*, Si Liu*, Zhen Miao, Fang Han, Raphael Gottardo, Wei Sun, Individual Level Differential Expression Analysis for Single Cell RNA-seq data, bioRxiv 2021.05.10.443350; doi: https://doi.org/10.1101/2021.05.10.443350  (* co-first author)
}
\keyword{ package }
